long for 32bit XChangeProperty
svn path=/trunk/kdelibs/; revision=382863

+++ kdecore/kwin.cpp
@@ -549,7 +549,7 @@ void KWin::setOpacity( WId win, uint percent )
         XDeleteProperty (qt_xdisplay(), win, kde_wm_window_opacity);
     else
     {
-        uint opacity = 0xFFFFFFFF/100.0*percent;
+        long opacity = 0xFFFFFFFF/100.0*percent;
         XChangeProperty(qt_xdisplay(), win, kde_wm_window_opacity, XA_CARDINAL, 32, PropModeReplace, (unsigned char *) &opacity, 1L);
     }
 #endif
@@ -558,7 +558,7 @@ void KWin::setOpacity( WId win, uint percent )
 void KWin::setShadowSize( WId win, uint percent )
 {
 #ifdef Q_QS_X11
-    uint shadowSize = 0xFFFFFFFF/100.0*percent;
+    long shadowSize = 0xFFFFFFFF/100.0*percent;
     XChangeProperty(qt_xdisplay(), win, kde_wm_window_shadow, XA_CARDINAL, 32, PropModeReplace, (unsigned char *) &shadowSize, 1L);
 #endif
 }
